requirements.txt

pandas                  
psycopg2-binary         
              


streamlit              
streamlit-authenticator 


google-genai                